daily_helper/requirements.txt
newsapi-python
google-genai
python-dotenv
